% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{linear_cov}
\alias{linear_cov}
\title{Linear kernel function}
\usage{
linear_cov(x1, x2, sigma = 0.3, sigma_b = 0.8, c = 0)
}
\arguments{
\item{x1}{\code{numeric} vector of values for first dataset}

\item{x2}{\code{numeric} vector of values for second dataset}

\item{sigma}{\code{numeric} scalar denoting hyperparameter for average distance from the mean. Defaults to \code{0.3}}

\item{sigma_b}{\code{numeric} scalar denoting hyperparameter for certainty around C. Defaults to \code{0.8}}

\item{c}{\code{numeric} scalar denoting hyperparameter for offset. Defaults to \code{0}}
}
\value{
\code{numeric} vector of covariance values
}
\description{
Linear kernel function
}
\author{
Trent Henderson
}
