% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{matern_cov}
\alias{matern_cov}
\title{Matern kernel function}
\usage{
matern_cov(x1, x2, sigma = 0.3, l = 0.8, nu = 1.5)
}
\arguments{
\item{x1}{\code{numeric} vector of values for first dataset}

\item{x2}{\code{numeric} vector of values for second dataset}

\item{sigma}{\code{numeric} scalar denoting hyperparameter for average distance from the mean. Defaults to \code{0.3}}

\item{l}{\code{numeric} scalar denoting hyperparameter for reach of influence on neighbouring points. Defaults to \code{0.8}}

\item{nu}{\code{numeric} scalar denoting hyperparameter for smoothness. Defaults to \code{1.5}}
}
\value{
\code{numeric} vector of covariance values
}
\description{
Matern kernel function
}
\author{
Trent Henderson
}
