% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{noise_cov}
\alias{noise_cov}
\title{White noise kernel function}
\usage{
noise_cov(x1, x2, sigma = 0.3)
}
\arguments{
\item{x1}{\code{numeric} vector of values for first dataset}

\item{x2}{\code{numeric} vector of values for second dataset}

\item{sigma}{\code{numeric} scalar denoting hyperparameter for average distance from the mean. Defaults to \code{0.3}}
}
\value{
\code{numeric} vector of covariance values
}
\description{
White noise kernel function
}
\author{
Trent Henderson
}
