% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{periodic_cov}
\alias{periodic_cov}
\title{Periodic kernel function}
\usage{
periodic_cov(x1, x2, sigma = 0.3, p = 0.8, l = 0.5)
}
\arguments{
\item{x1}{\code{numeric} vector of values for first dataset}

\item{x2}{\code{numeric} vector of values for second dataset}

\item{sigma}{\code{numeric} scalar denoting hyperparameter for average distance from the mean. Defaults to \code{0.3}}

\item{p}{\code{numeric} scalar denoting hyperparameter for distance between repetitions. Defaults to \code{0.8}}

\item{l}{\code{numeric} scalar denoting hyperparameter for reach of influence on neighbouring points. Defaults to \code{0.5}}
}
\value{
\code{numeric} vector of covariance values
}
\description{
Periodic kernel function
}
\author{
Trent Henderson
}
