% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample from GP posterior using the posterior mean vector and covariance matrix}
\usage{
sample_posterior(gp, n = 5, seed = 123)
}
\arguments{
\item{gp}{\code{GP} object containing the posterior mean vector and covariance matrix}

\item{n}{\code{integer} denoting the number of samples to draw. Defaults to \code{5}}

\item{seed}{\code{integer} denoting a fix for R's pseudorandom number generator. Defaults to \code{123}}
}
\value{
\code{matrix} of samples
}
\description{
Sample from GP posterior using the posterior mean vector and covariance matrix
}
\author{
Trent Henderson
}
