% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{se_cov}
\alias{se_cov}
\title{Squared exponential kernel function (radial basis function)}
\usage{
se_cov(x1, x2, sigma = 0.3, l = 0.8)
}
\arguments{
\item{x1}{\code{numeric} vector of values for first dataset}

\item{x2}{\code{numeric} vector of values for second dataset}

\item{sigma}{\code{numeric} scalar denoting hyperparameter for average distance from the mean. Defaults to \code{0.3}}

\item{l}{\code{numeric} scalar denoting hyperparameter for reach of influence on neighbouring points. Defaults to \code{0.8}}
}
\value{
\code{numeric} vector of covariance values
}
\description{
Squared exponential kernel function (radial basis function)
}
\author{
Trent Henderson
}
